% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visumot_all.R
\name{visumot_all}
\alias{visumot_all}
\title{visumot_all}
\usage{
visumot_all(
  df,
  images,
  frame_range = NULL,
  display_frame = TRUE,
  display_summary = FALSE,
  visumot_frame.list = NULL,
  visumot_summary.list = NULL,
  save = TRUE,
  file_name = NULL,
  width = 2000,
  height = 1200,
  rel_width = 1/3,
  browse = FALSE,
  return = FALSE
)
}
\arguments{
\item{df}{dataframe of the form: \code{df(track, time, X, Y, mapping_parameters, ...)}}

\item{images}{\code{vector}: filenames of images ordered by frames/time}

\item{frame_range}{\code{integer}: frames to be mapped; default=\code{NULL} maps full time/frame range found in \code{df}}

\item{display_frame}{\code{logical}: display output from \code{\link{visumot_frame}}}

\item{display_summary}{\code{logical}: display output from \code{\link{visumot_summary}}}

\item{visumot_frame.list}{\code{list}: parameters passed to \code{\link{visumot_frame}}}

\item{visumot_summary.list}{\code{list}: parameters passed to \code{\link{visumot_summary}}}

\item{save}{\code{logical}: saves tiff-stack or image to ./output/visumotr-date.tiff}

\item{width}{\code{numeric}: width of output tiff in pixels}

\item{height}{\code{numeric}: height of output tiff in pixels}

\item{rel_width}{\code{numeric}: \code{if(display_frame & display_summary)}, \code{rel_width} defines
the relative width of the summary plot to the mapped frame}

\item{browse}{\code{logical}: browse output with default system application for tiffs, e.g. imageJ}

\item{return}{\code{logical}: return images-series as magick image-stack}
}
\description{
Creates time-resolved images series of tracking data with modifiable parameter mapping. 
\code{visumot_all} is a wrapper function for \code{\link{visumot_frame}} and \code{\link{visumot_summary}}. User parameters for both
functions can be passed via the lists \code{visumot_frame.list} and \code{visumot_summary.list}.
}
\details{

}
\examples{
# import tracking data
df <- read.csv('hiv_tracking.csv')
# get image files
images <- list.files(pattern='.tif')
# run visumotR with default settings
visumot_all(df, images)
# run visumotR with visumot_frame and visumot_summary
tracks <- c(34, 125, 199, 205)
summary_pars <- list(par.map='angle', time.unit='min', tracks=tracks)
frame_pars <- list(par.map='angle', par.shape='infected', tracks=tracks)
visumot_all(df,
 images,
 frame_range = c(20,100),
 display_summary = 1,
 visumot_frame.list = frame_pars,
 visumot_summary.list = summary_pars,
 file_name = 'hiv_tracking_angle'
 )
}
