% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visumot_frame.R
\name{visumot_frame}
\alias{visumot_frame}
\title{visumot_frame}
\usage{
visumot_frame(df, ..., all.list = FALSE)
}
\arguments{
\item{df}{dataframe of the form: \code{df(track, time, X, Y, mapping_parameters, ...)}}

\item{image}{\code{character}: filename of image}

\item{frame}{\code{integer}: frame to be mapped}

\item{tracks}{\code{vector}: defining tracks to be displayed}

\item{par.map}{\code{character}: specifying parameter in \code{df} to be visualized by color}

\item{par.shape}{\code{character}: specifying parameter in \code{df} to be mapped on shape}

\item{par.display}{display option for mapping; default: \code{TRUE}, mapping is disable with: \code{FALSE}}

\item{par.max}{\code{numeric}: defining range of color mapping}

\item{par.unit}{\code{character}: unit of the numeric mapped parameter}

\item{crop}{\code{logical} option for cropping images; default: \code{FALSE}}

\item{sub.img}{\code{logical} option for creating sub-images from specified \code{tracks} or pre-filtered \code{df}; default: \code{FALSE}}

\item{sub.window}{\code{numeric}: size of the sub-images in pixels}

\item{sub.col}{\code{numeric}: number of columns in which sub-images are arranged}

\item{tracks.size}{\code{numeric}: size of tracks}

\item{tracks.alpha}{\code{numeric}: transparency of tracks}

\item{tracks.length}{\code{numeric}: length of tracks (in frames)}

\item{points.size}{\code{numeric}: size of points}

\item{points.alpha}{\code{numeric}: transparency of points}

\item{points.stat}{\code{character}: display statistic; default: \code{'echo'}, for blurring; without blurring \code{'identity'}}

\item{points.shape}{\code{numeric}: set shape from ggplot2 shape palette}

\item{axis.tick}{\code{numeric}: axis ticks in px}

\item{axis.display}{\code{logical}: display axis}

\item{axis.labs}{\code{logical}: display labs}

\item{unit}{\code{character}: setting name of unit; default: \code{'px'}}

\item{scaling}{\code{numeric}: scaling factor for unit; default: \code{1}}

\item{scale.bar}{\code{logical}: show scalebar; default: \code{FALSE}}

\item{scale.width}{\code{numeric}: width of scalebar; default: \code{40}}

\item{scale.height}{\code{numeric}: height of scalebar; default: \code{10}}

\item{scale.x}{\code{numeric}: distance from left border of the image towards scalebar}

\item{scale.y}{\code{numeric}: distance from bottom border of the image towards scalebar}

\item{scale.color}{\code{character}: specify color from R-color palette or hexcode}
}
\value{
ggplot2 plot-object
}
\description{
To be written...
}
\details{
To be written...
}
\examples{
visumot_frame()
}
